+++ CHANGELOG.md
@@ -66,6 +66,7 @@ Breaking changes will be documented in this changelog file for each version.
 * github-366: Feature parity between Flowman shell and Flowman remote shell
 * github-365: Implement saving mappings in Flowman Kernel/client
 * github-367: Create integration test for "quickstart" archetype
+* github-375: "project reload" does not work correctly in remote shell with nested directories
 
 ### Breaking changes